requirements.txt
click==7.1.2
PyYAML==5.4.1
appdirs==1.4.4
click==7.1.2
tabulate==0.8.9
yaspin==1.5.0
requests==2.25.1
